+++ migrations/ordinals/V8__counts_by_block.sql
@@ -1,5 +1,5 @@
 CREATE TABLE counts_by_block (
-    block_height BIGINT NOT NULL PRIMARY KEY,
+    block_height NUMERIC NOT NULL PRIMARY KEY,
     block_hash TEXT NOT NULL,
     inscription_count BIGINT NOT NULL,
     inscription_count_accum BIGINT NOT NULL,
